requirements.txt
streamlit>=1.40.0
openai>=1.30.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
